+++ src/js/media-modal.js
@@ -1,0 +1,7 @@
+/**
+ * Media modal entry stub for @wordpress/scripts build.
+ *
+ * @package IMGVerse
+ */
+
+console.log( 'imgverse media-modal stub' );